addressing review comments
- moved enum states to subsystem
- renamed commands to match states
- made things more organized✨
- also deleted sensor in transfer cuz alice already coded one in main :3
