+++ src/main/java/frc/robot/commands/States/Ejecting.java
@@ -7,18 +7,18 @@ package frc.robot.commands.States;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants.constIntake;
 import frc.robot.Constants.constTransfer;
-import frc.robot.RobotContainer.RobotState;
+import frc.robot.subsystems.StateMachine.RobotState;
+import frc.robot.subsystems.Intake;
 import frc.robot.subsystems.StateMachine;
 import frc.robot.subsystems.Transfer;
-import frc.robot.subsystems.Intake;
 
-public class IntakeFloor extends Command {
+public class Ejecting extends Command {
   StateMachine subStateMachine;
   Intake subIntake;
   Transfer subTransfer;
 
-  /** Creates a new Intake. */
-  public IntakeFloor(StateMachine subStateMachine, Intake subIntake, Transfer subTransfer) {
+  /** Creates a new Eject. */
+  public Ejecting(StateMachine subStateMachine, Intake subIntake, Transfer subTransfer) {
     this.subStateMachine = subStateMachine;
     this.subIntake = subIntake;
     this.subTransfer = subTransfer;
@@ -30,9 +30,9 @@ public class IntakeFloor extends Command {
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    subStateMachine.setRobotState(RobotState.INTAKING);
-    subIntake.setIntakeRollerSpeed(constIntake.INTAKING_SPEED);
-    subTransfer.setFeederSpeed(constTransfer.INTAKING_SPEED);
+    subStateMachine.setRobotState(RobotState.EJECTING);
+    subIntake.setIntakeRollerSpeed(constIntake.EJECTING_SPEED);
+    subTransfer.setFeederSpeed(constTransfer.EJECTING_SPEED);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
@@ -43,14 +43,12 @@ public class IntakeFloor extends Command {
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    if (interrupted) {
-      subStateMachine.setRobotState(RobotState.NONE);
-    }
+    subStateMachine.setRobotState(RobotState.NONE);
   }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return subTransfer.isGamePieceCollected();
+    return false;
   }
 }
