+++ src/main/java/frc/robot/commands/States/NoneState.java
@@ -7,7 +7,7 @@ package frc.robot.commands.States;
 import edu.wpi.first.units.Units;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants.constElevator;
-import frc.robot.RobotContainer.RobotState;
+import frc.robot.subsystems.StateMachine.RobotState;
 import frc.robot.subsystems.Elevator;
 import frc.robot.subsystems.Intake;
 import frc.robot.subsystems.Shooter;
